requirements.txt
black==23.10.1
click==8.1.7
greenlet==3.0.1
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
protobuf==4.21.12
SQLAlchemy==2.0.23
typing_extensions==4.8.0
